u2f: dropped old implementation
